+++ R/geom_treemap_subgroup_text.R
@@ -15,10 +15,10 @@
 #' here. For full details on how these options work, see the documentation for
 #' `ggfittext::geom_fit_text`.
 #'
-#' The `fixed` argument is used to set the treemap layout algorithm. All
-#' `treemapify` geoms added to a plot should have the same value for `fixed`,
-#' or they will not share a common layout (see `geom_treemap` for details on
-#' the layout algorithms).
+#' The `layout` argument is used to set the treemap layout algorithm. All
+#' `treemapify` geoms added to a plot should have the same value for `layout`
+#' and `start`, or they will not share a common layout (see `geom_treemap` for
+#' details on the layout algorithms).
 #'
 #' @param padding.x,padding.y `grid::unit` object, giving horizontal or vertical
 #' padding between text and edge of tile. Defaults to 1 mm.
@@ -30,14 +30,19 @@
 #' @param grow If `TRUE`, text will be grown as well as shrunk to fill the box.
 #' @param reflow If `TRUE`, text will be reflowed (wrapped) to better fit the
 #' box.
-#' @param fixed If `TRUE`, the alternative 'fixed' tile layout algorithm will be
-#' used.
+#' @param layout The layout algorithm, one of either "squarified" (the
+#' default), "scol", "srow" or "fixed". See Details for full details on the
+#' different layout algorithms.
+#' @param start The corner in which to start placing the tiles. One of
+#' 'bottomleft' (the default), 'topleft', 'topright' or 'bottomright'.
 #' @param mapping,data,stat,position,na.rm,show.legend,inherit.aes,... Standard
 #' geom arguments as for `ggplot2::geom_text`.
 #' @param level One of 'subgroup', 'subgroup2' or 'subgroup3', giving the
 #' subgrouping level for which to draw text labels. It is recommended to use
 #' the aliases `geom_treemap_subgroup2_text` and `geom_treemap_subgroup3_text`
 #' instead of this argument.
+#' @param fixed Deprecated. Use `layout = "fixed"` instead. Will be removed in
+#' later versions.
 #'
 #' @seealso geom_treemap, geom_treemap_subgroup_border
 #'
@@ -78,7 +83,9 @@ geom_treemap_subgroup_text <- function(
   min.size = 4,
   grow = FALSE,
   reflow = FALSE,
-  fixed = FALSE,
+  fixed = NULL,
+  layout = "squarified",
+  start = "bottomleft",
   level = "subgroup",
   ...
 ) {
@@ -99,6 +106,8 @@ geom_treemap_subgroup_text <- function(
       grow = grow,
       reflow = reflow,
       fixed = fixed,
+      layout = layout,
+      start = start,
       level = level,
       ...
     )
@@ -133,7 +142,9 @@ GeomSubgroupText <- ggplot2::ggproto(
     grow = FALSE,
     reflow = FALSE,
     place = "bottom",
-    fixed = FALSE,
+    fixed = NULL,
+    layout = "squarified",
+    start = "bottomleft",
     level = "subgroup"
   ) {
 
@@ -179,7 +190,9 @@ GeomSubgroupText <- ggplot2::ggproto(
     params <- list(
       data = data,
       area = "area",
-      fixed = fixed
+      fixed = fixed,
+      layout = layout,
+      start = start
     )
     for (l in levels[1:(length(levels) - 1)]) { params[l] <- l }
     data <- do.call(treemapify, params)