+++ R/geom_treemap_text.R
@@ -11,8 +11,9 @@
 #' reflowing, etc.) are also available here. For full details on how these
 #' options work, see the documentation for `ggfittext::geom_fit_text`.
 #'
-#' All `treemapify` geoms added to a plot should have the same value for
-#' `fixed`, or they will not share a common layout (see `geom_treemap` for
+#' The `layout` argument is used to set the treemap layout algorithm. All
+#' `treemapify` geoms added to a plot should have the same value for `layout`
+#' and `start`, or they will not share a common layout (see `geom_treemap` for
 #' details on the layout algorithms).
 #'
 #' @param padding.x,padding.y `grid::unit` object, giving horizontal or vertical
@@ -25,10 +26,15 @@
 #' @param grow If `TRUE`, text will be grown as well as shrunk to fill the box.
 #' @param reflow If `TRUE`, text will be reflowed (wrapped) to better fit the
 #' box.
-#' @param fixed If `TRUE`, the alternative 'fixed' tile layout algorithm will be
-#' used.
+#' @param layout The layout algorithm, one of either "squarified" (the
+#' default), "scol", "srow" or "fixed". See Details for full details on the
+#' different layout algorithms.
+#' @param start The corner in which to start placing the tiles. One of
+#' 'bottomleft' (the default), 'topleft', 'topright' or 'bottomright'.
 #' @param mapping,data,stat,position,na.rm,show.legend,inherit.aes,... Standard
 #' geom arguments as for `ggplot2::geom_text`.
+#' @param fixed Deprecated. Use `layout = "fixed"` instead. Will be removed in
+#' later versions.
 #'
 #' @seealso geom_treemap
 #'
@@ -69,7 +75,9 @@ geom_treemap_text <- function(
   min.size = 4,
   grow = FALSE,
   reflow = FALSE,
-  fixed = FALSE,
+  fixed = NULL,
+  layout = "squarified",
+  start = "bottomleft",
   ...
 ) {
   ggplot2::layer(
@@ -89,6 +97,8 @@ geom_treemap_text <- function(
       grow = grow,
       reflow = reflow,
       fixed = fixed,
+      layout = layout,
+      start = start,
       ...
     )
   )
@@ -121,7 +131,9 @@ GeomTreemapText <- ggplot2::ggproto(
     min.size = 4,
     grow = FALSE,
     reflow = FALSE,
-    fixed = FALSE,
+    fixed = NULL,
+    layout = "squarified",
+    start = "bottomleft",
     place = "centre"
   ) {
 
@@ -131,7 +143,9 @@ GeomTreemapText <- ggplot2::ggproto(
     params <- list(
       data = data,
       area = "area",
-      fixed = fixed
+      fixed = fixed,
+      layout = layout,
+      start = start
     )
     for (subgrouplevel in c("subgroup", "subgroup2", "subgroup3")) {
       if (subgrouplevel %in% names(data)) {